+++ zowex/src/main.rs
@@ -9,7 +9,7 @@
 *
 */
 
-use std::collections::HashMap;
+use serde::{Deserialize, Serialize};
 use std::env;
 use std::ffi::OsString;
 use std::io::prelude::*;
@@ -31,18 +31,9 @@ use rpassword::read_password;
 extern crate sysinfo;
 use sysinfo::{ProcessExt, System, SystemExt};
 
-
-// NOTE(Kelosky): these sync with imperative header values
-const X_ZOWE_DAEMON_HEADERS: &str = "x-zowe-daemon-headers";
-const X_ZOWE_DAEMON_EXIT: &str = "x-zowe-daemon-exit";
-const X_ZOWE_DAEMON_END: &str = "x-zowe-daemon-end";
-const X_ZOWE_DAEMON_PROGRESS: &str = "x-zowe-daemon-progress";
-const X_ZOWE_DAEMON_PROMPT: &str = "x-zowe-daemon-prompt";
-const X_ZOWE_DAEMON_VERSION: &str = "x-zowe-daemon-version";
-const X_HEADERS_VERSION_ONE_LENGTH: usize = 8;
 const DEFAULT_PORT: i32 = 4000;
 
-const X_ZOWE_DAEMON_REPLY: &str = "x-zowe-daemon-reply:";
+const X_ZOWE_DAEMON_REPLY: &str = "daemon-client";
 
 const EXIT_CODE_CANNOT_CONNECT_TO_RUNNING_DAEMON: i32 = 100;
 const EXIT_CODE_CANNOT_GET_MY_PATH: i32 = 101;
@@ -56,7 +47,24 @@ struct DaemonProcInfo {
     is_running: bool,
     name: String,
     pid: String,
-    cmd: String
+    cmd: String,
+}
+
+#[derive(Serialize, Deserialize)]
+#[allow(non_snake_case)]
+struct DaemonRequest {
+    stdout: Option<String>,
+    stderr: Option<String>,
+    exitCode: Option<i32>,
+    progress: Option<bool>,
+    prompt: Option<String>,
+    securePrompt: Option<String>,
+}
+
+#[derive(Serialize, Deserialize)]
+struct DaemonResponse {
+    reply: String,
+    id: String,
 }
 
 // TODO(Kelosky): performance tests, `time for i in {1..10}; do zowe -h >/dev/null; done`
@@ -92,9 +100,12 @@ fn main() -> std::io::Result<()> {
                  * the command run by the daemon.
                  */
                 cmd_result = Ok(0);
-            },
+            }
             Err(error) => {
-                println!("The daemon failed to run your command due to this error:\n{}", error);
+                println!(
+                    "The daemon failed to run your command due to this error:\n{}",
+                    error
+                );
                 cmd_result = Ok(EXIT_CODE_DAEMON_FAILED_TO_RUN_CMD);
             }
         }
@@ -107,9 +118,12 @@ fn main() -> std::io::Result<()> {
     match cmd_result {
         Ok(value) => {
             std::process::exit(value);
-        },
+        }
         Err(error) => {
-            println!("NodeJS zowe failed to run your command due to this error:\n{}", error);
+            println!(
+                "NodeJS zowe failed to run your command due to this error:\n{}",
+                error
+            );
             std::process::exit(EXIT_CODE_FAILED_TO_RUN_NODEJS_CMD);
         }
     }
@@ -162,15 +176,20 @@ fn run_daemon_command(mut args: String) -> std::io::Result<()> {
     // form our host, port, and connection strings
     let daemon_host = "127.0.0.1".to_owned();
     let port_string = get_port_string();
-    let host_port_conn_str = format!("{}:{}", daemon_host, port_string);
 
+    let mut stream = establish_connection(daemon_host, port_string)?;
+    Ok(talk(&_resp, &mut stream)?)
+}
+
+fn establish_connection(host: String, port: String) -> std::io::Result<TcpStream> {
     /* Attempt to make a TCP connection to the daemon.
      * Iterate to enable a slow system to start the daemon.
      */
+    let host_port_conn_str = format!("{}:{}", host, port);
     let mut conn_attempt = 1;
     let mut we_started_daemon = false;
     let mut cmd_to_show: String = String::new();
-    let mut stream = loop {
+    let stream = loop {
         let conn_result = TcpStream::connect(&host_port_conn_str);
         if let Ok(good_stream) = conn_result {
             // We made our connection. Break with the actual stream value
@@ -187,12 +206,13 @@ fn run_daemon_command(mut args: String) -> std::io::Result<()> {
                 let njs_zowe_path = get_nodejs_zowe_path();
                 we_started_daemon = true;
                 cmd_to_show = start_daemon(&njs_zowe_path);
-    } else {
+            } else {
                 if we_started_daemon {
                     println!("The Zowe daemon that we started is not running on host = {} with port = {}.",
-                        daemon_host, port_string
+                        host, port
                     );
-                    println!("Command used to start the Zowe daemon was:\n    {}\nTerminating.",
+                    println!(
+                        "Command used to start the Zowe daemon was:\n    {}\nTerminating.",
                         cmd_to_show
                     );
                     std::process::exit(EXIT_CODE_DEAMON_NOT_RUNNING_AFTER_START);
@@ -202,10 +222,13 @@ fn run_daemon_command(mut args: String) -> std::io::Result<()> {
 
         if conn_attempt == 5 {
             println!("\nUnable to connect to Zowe daemon with name = {} and pid = {} on host = {} and port = {}.",
-                daemon_proc_info.name, daemon_proc_info.pid, daemon_host, port_string
+                daemon_proc_info.name, daemon_proc_info.pid, host, port
             );
-            println!("Command = {}\nTerminating after maximum retries.", daemon_proc_info.cmd);
-             std::process::exit(EXIT_CODE_CANNOT_CONNECT_TO_RUNNING_DAEMON);
+            println!(
+                "Command = {}\nTerminating after maximum retries.",
+                daemon_proc_info.cmd
+            );
+            std::process::exit(EXIT_CODE_CANNOT_CONNECT_TO_RUNNING_DAEMON);
         }
 
         // pause between attempts to connect
@@ -216,94 +239,106 @@ fn run_daemon_command(mut args: String) -> std::io::Result<()> {
         conn_attempt = conn_attempt + 1;
     };
 
-    stream.write(_resp).unwrap(); // write it
+    Ok(stream)
+}
+
+fn talk(message: &[u8], stream: &mut TcpStream) -> std::io::Result<()> {
+    /*
+     * Send the command line arguments to the daemon and await responses.
+     */
+    stream.write(message).unwrap(); // write it
     let mut stream_clone = stream.try_clone().expect("clone failed");
 
-    let mut reader = BufReader::new(&mut stream);
-    let mut headers: HashMap<String, i32> = HashMap::new();
+    let mut reader = BufReader::new(&*stream);
+
+    let mut exit_code = 0;
+    let mut _progress = false;
 
     loop {
-        let mut line = String::new();
+        let mut u_payload: Vec<u8> = Vec::new();
+        let payload: String;
 
-        if reader.read_line(&mut line).unwrap() > 0 {
-            // first break if identified headers exist on it
-            let pieces = get_beg(&line);
-            let new_headers: HashMap<std::string::String, i32>;
+        // read until form feed (\f)
+        if reader.read_until(0xC, &mut u_payload).unwrap() > 0 {
+            // remove form feed and convert to a string
+            u_payload.pop(); // remove the 0xC
+            payload = str::from_utf8(&u_payload).unwrap().to_string();
 
-            // println!("@debug {}", line);
-
-            // if headers include other data on the same line, isolate them
-            if pieces.len() > 1 {
-                new_headers = parse_headers(&pieces[0]);
-            // println!("@debug data on pieces");
-            // otherwise, get raw headers
-            } else {
-                new_headers = parse_headers(&line);
-            }
-            let mut got_new_headers = false;
-
-            // adjust so that `headers` always contains the last sent headers
-            if new_headers.len() > 0 {
-                headers = new_headers;
-                got_new_headers = true;
-            // println!("@debug got new headers");
-            } else {
-                // do nothing
-                // println!("@debug did not got new headers");
-            }
-
-            // if no headers, print the line as it comes in
-            // TODO(Kelosky): if stderr, print stderr
-            if headers.len() == 0 {
-                print!("{}", line);
-                io::stdout().flush().unwrap();
-            } else {
-                // we have headers but this statement that we read does not contain new header values
-                if got_new_headers == false {
-                    let &progress = headers.get(X_ZOWE_DAEMON_PROGRESS).unwrap();
-
-                    // if progress bar is in place, strip off the newline character
-                    if progress == 1i32 {
-                        print!("{}", &line[0..(line.len() - 1)]);
-                        io::stdout().flush().unwrap();
-                    } else {
-                        print!("{}", line);
-                        io::stdout().flush().unwrap();
-                    }
-                // else, we received headers and may need to print extraneous data on the same line
-                } else {
-                    if pieces.len() > 1 {
-                        print!("{}", pieces[1]);
-                        io::stdout().flush().unwrap();
-                        let &prompt = headers.get(X_ZOWE_DAEMON_PROMPT).unwrap();
-
-                        if prompt > 0i32 {
-                            // prompt
-                            let mut reply = String::new();
-
-                            // type 2 indicates secure fields
-                            if prompt == 2i32 {
-                                reply = read_password().unwrap();
-                            // else regular prompting
-                            } else {
-                                io::stdin().read_line(&mut reply).unwrap();
-                            }
-
-                            // append response to header
-                            let mut full_reply = X_ZOWE_DAEMON_REPLY.to_owned();
-                            full_reply.push_str(&reply);
-
-                            // adjust our state that prompting has been resolved
-                            headers.insert(X_ZOWE_DAEMON_PROMPT.to_string(), 0i32);
-
-                            // write response
-                            stream_clone.write(full_reply.as_bytes()).unwrap();
-
-                            // else, just write this line
-                        }
+            let p: DaemonRequest = match serde_json::from_str(&payload) {
+                Ok(p) => p,
+                Err(_e) => {
+                    // TODO(Kelosky): handle this only if progress bar mode is active
+                    print!("{}", payload);
+                    io::stdout().flush().unwrap();
+                    DaemonRequest {
+                        stdout: None,
+                        stderr: None,
+                        exitCode: Some(0i32),
+                        progress: None,
+                        prompt: None,
+                        securePrompt: None,
                     }
                 }
+            };
+
+            match p.stdout {
+                Some(s) => {
+                    print!("{}", s);
+                    io::stdout().flush().unwrap();
+                }
+                None => (), // do nothing
             }
+
+            match p.stderr {
+                Some(s) => {
+                    eprint!("{}", s);
+                    io::stdout().flush().unwrap();
+                }
+                None => (), // do nothing
+            }
+
+            match p.prompt {
+                Some(s) => {
+                    print!("{}", s);
+                    io::stdout().flush().unwrap();
+                    let mut reply = String::new();
+                    io::stdin().read_line(&mut reply).unwrap();
+                    let response: DaemonResponse = DaemonResponse {
+                        reply,
+                        id: X_ZOWE_DAEMON_REPLY.to_string(),
+                    };
+                    let v = serde_json::to_string(&response)?;
+
+                    stream_clone.write(v.as_bytes()).unwrap();
+                }
+                None => (), // do nothing
+            }
+
+            match p.securePrompt {
+                Some(s) => {
+                    print!("{}", s);
+                    io::stdout().flush().unwrap();
+                    let reply;
+                    reply = read_password().unwrap();
+                    let response: DaemonResponse = DaemonResponse {
+                        reply,
+                        id: X_ZOWE_DAEMON_REPLY.to_string(),
+                    };
+                    let v = serde_json::to_string(&response)?;
+                    stream_clone.write(v.as_bytes()).unwrap();
+                }
+                None => (), // do nothing
+            }
+
+            exit_code = match p.exitCode {
+                Some(s) => s,
+                None => 0, // do nothing
+            };
+
+            _progress = match p.progress {
+                Some(s) => s,
+                None => false, // do nothing
+            };
         } else {
             // end of reading
             break;
@@ -312,96 +347,24 @@ fn run_daemon_command(mut args: String) -> std::io::Result<()> {
 
     stream.shutdown(Shutdown::Both)?; // terminate connection
 
-    if headers.contains_key(X_ZOWE_DAEMON_EXIT) {
-        let &exit = headers.get(X_ZOWE_DAEMON_EXIT).unwrap();
-        if exit == 1i32 {
-            std::process::exit(exit);
-        }
+    // TODO(Kelosky): maybe this should just be a `return Err`
+    if exit_code != 0 {
+        std::process::exit(exit_code);
     }
 
     Ok(())
-}
-
-fn parse_headers(buf: &String) -> HashMap<String, i32> {
-    let mut headers: HashMap<String, i32> = HashMap::new();
-
-    // break response on newline
-    let mut lines = buf.lines();
-    let first = lines.next().unwrap();
-
-    let raw_headers: Vec<&str> = first.split(';').collect();
-    // must match minimum length
-    if raw_headers.len() >= X_HEADERS_VERSION_ONE_LENGTH {
-        // first and last headers must be as expected
-        if raw_headers[0].contains(X_ZOWE_DAEMON_HEADERS)
-            && raw_headers[X_HEADERS_VERSION_ONE_LENGTH - 1].contains(X_ZOWE_DAEMON_END)
-        {
-            for raw_header in raw_headers.iter() {
-                let parts: Vec<&str> = raw_header.split(':').collect();
-                let key = parts[0].to_owned();
-                let char_value = parts[1].trim();
-                let int_value = char_value.parse::<i32>().unwrap();
-                headers.insert(key, int_value);
-            }
-        }
-    }
-
-    // we have a version header
-    if headers.contains_key(X_ZOWE_DAEMON_VERSION) {
-        let &version = headers.get(X_ZOWE_DAEMON_VERSION).unwrap();
-
-        // version is not an int
-        if version != 1i32 {
-            headers.clear();
-        }
-    // else, new version, clear data
-    } else {
-        headers.clear();
-    }
-
-    headers
 }
 
 fn get_port_string() -> String {
     let mut _port = DEFAULT_PORT;
 
     match env::var("ZOWE_DAEMON") {
+        // TODO(Kelosky): handle unwrap properly
         Ok(val) => _port = val.parse::<i32>().unwrap(),
         Err(_e) => _port = DEFAULT_PORT,
     }
     let port_string = _port.to_string();
     return port_string;
-}
-
-fn get_beg(buf: &str) -> Vec<String> {
-    let mut parts: Vec<String> = Vec::new();
-
-    let ss: Vec<char> = buf.chars().collect();
-
-    // if this line is long enough to contain headers
-    if ss.len() >= 2 {
-        // if this line begins with header info, just return and parse it.  there is no extraneous data at the beginning
-        if ss[0] == 'x' && ss[1] == '-' {
-            parts.push(buf.to_owned());
-        // else, if we find we have some data in the string and a header, e.g. `}x-zowe-...` parse of the beginning data and header
-        // into two pieces to pass back
-        } else {
-            for (i, x) in ss.iter().enumerate() {
-                if *x == 'x' && ss[i + 1] == '-' {
-                    let first: String = ss.iter().skip(0).take(i).collect();
-                    let second: String = ss.iter().skip(i).take(ss.len() - 1).collect();
-                    parts.push(second);
-                    parts.push(first);
-                    break;
-                }
-            }
-        }
-    // line cannot have headers, just return the line
-    } else {
-        parts.push(buf.to_owned());
-    }
-
-    return parts;
 }
 
 // Get the file path to the command that runs the NodeJS version of Zowe
@@ -478,7 +441,7 @@ fn is_daemon_running() -> DaemonProcInfo {
                 is_running: true,
                 name: process.name().to_string(),
                 pid: pid.to_string(),
-                cmd: proc_cmd
+                cmd: proc_cmd,
             };
         }
     }
@@ -486,7 +449,7 @@ fn is_daemon_running() -> DaemonProcInfo {
         is_running: false,
         name: "no name".to_string(),
         pid: "no pid".to_string(),
-        cmd: "no cmd".to_string()
+        cmd: "no cmd".to_string(),
     };
 }
 
@@ -502,10 +465,11 @@ fn user_wants_daemon() -> bool {
         Err(_e) => env_var_val = "NoDaemon".to_string(),
     }
 
-    if env_var_val.to_lowercase() == "false" || env_var_val.to_lowercase() == "no" ||
-       env_var_val == "0"
+    if env_var_val.to_lowercase() == "false"
+        || env_var_val.to_lowercase() == "no"
+        || env_var_val == "0"
     {
-        return false
+        return false;
     }
     return true;
 }
@@ -531,10 +495,13 @@ fn run_nodejs_command(cmd_line_args: &mut Vec<String>) -> Result<i32, i32> {
     {
         Ok(new_proc) => {
             exit_code = new_proc.status.code().unwrap();
-        },
+        }
         Err(error) => {
             println!("Failed to run the following command:");
-            println!("    Program = {}\n    arguments = {:?}", njs_zowe_path, cmd_line_args);
+            println!(
+                "    Program = {}\n    arguments = {:?}",
+                njs_zowe_path, cmd_line_args
+            );
             println!("Due to this error:\n    {}", error);
             exit_code = EXIT_CODE_FAILED_TO_RUN_NODEJS_CMD;
         }
@@ -598,9 +565,12 @@ fn start_daemon(njs_zowe_path: &str) -> String {
         .stderr(Stdio::null())
         .spawn()
     {
-        Ok(_unused) => { /* nothing to do */ },
+        Ok(_unused) => { /* nothing to do */ }
         Err(error) => {
-            println!("Failed to start the following process:\n    {} {}", njs_zowe_path, daemon_arg);
+            println!(
+                "Failed to start the following process:\n    {} {}",
+                njs_zowe_path, daemon_arg
+            );
             println!("Due to this error:\n    {}", error);
             std::process::exit(EXIT_CODE_CANNOT_START_DAEMON);
         }
@@ -625,50 +595,7 @@ mod tests {
 
     #[test]
     fn test_parse_headers_get_total_length() {
-        let headers = "x-zowe-daemon-headers:8;x-zowe-daemon-version:1;x-zowe-daemon-exit:1;x-zowe-daemon-stdout:1;x-zowe-daemon-stderr:0;x-zowe-daemon-prompt:0;x-zowe-daemon-progress:0;x-zowe-daemon-end:0".to_string();
-        let headers_map = parse_headers(&headers);
-
-        // expect 8 entries in the map
-        assert_eq!(8, headers_map.len());
-    }
-
-    #[test]
-    fn test_parse_headers_get_length() {
-        let headers = "x-zowe-daemon-headers:8;x-zowe-daemon-version:1;x-zowe-daemon-exit:1;x-zowe-daemon-stdout:1;x-zowe-daemon-stderr:0;x-zowe-daemon-prompt:0;x-zowe-daemon-progress:0;x-zowe-daemon-end:0".to_string();
-        let headers_map = parse_headers(&headers);
-
-        let &len = headers_map
-            .get(&("x-zowe-daemon-headers".to_string()))
-            .unwrap();
-
-        // expect len code set from header
-        assert_eq!(8, len);
-    }
-
-    #[test]
-    fn test_parse_headers_get_version() {
-        let headers = "x-zowe-daemon-headers:8;x-zowe-daemon-version:1;x-zowe-daemon-exit:1;x-zowe-daemon-stdout:1;x-zowe-daemon-stderr:0;x-zowe-daemon-prompt:0;x-zowe-daemon-progress:0;x-zowe-daemon-end:0".to_string();
-        let headers_map = parse_headers(&headers);
-
-        let &len = headers_map
-            .get(&("x-zowe-daemon-version".to_string()))
-            .unwrap();
-
-        // expect version code set from header
-        assert_eq!(1, len);
-    }
-
-    #[test]
-    fn test_parse_headers_get_exit() {
-        let headers = "x-zowe-daemon-headers:8;x-zowe-daemon-version:1;x-zowe-daemon-exit:1;x-zowe-daemon-stdout:1;x-zowe-daemon-stderr:0;x-zowe-daemon-prompt:0;x-zowe-daemon-progress:0;x-zowe-daemon-end:0".to_string();
-        let headers_map = parse_headers(&headers);
-
-        let &exit = headers_map
-            .get(&("x-zowe-daemon-exit".to_string()))
-            .unwrap();
-
-        // expect exit code set from header
-        assert_eq!(1, exit);
+        assert_eq!(8, 8);
     }
 
     #[test]